requirements.txt
mediapipe
python-dotenv
opencv-python
gradio
numpy
imageio[ffmpeg]
ultralytics>=8.3.0
roboflow
inference
